+++ src/component/bottomBar/bar/DotColor.jsx
@@ -45,8 +45,6 @@ const DotColor = React.memo(() => {
         setOpen(true);
     }
 
-    console.log("dotColor");
-
     return (
         <Main open={open}>
             <Element