+++ src/component/bottomBar/storage/Storage.jsx
@@ -13,9 +13,6 @@ export default React.memo((
     }) => {
     const { Juuten_Storage } = useSelector(selectCollection)
 
-    console.log(Juuten_Storage);
-
-    console.log('storage');
     return (
         <BottemBarTemplate
             useContainer={false}