+++ src/component/folder/FolderBlock.jsx
@@ -53,8 +53,6 @@ export default React.memo((
     const navigate = useNavigate()
     const dispatch = useDispatch()
 
-    console.log('folderBlock')
-
     function editFolder(e, item) {
         e.stopPropagation()
         dispatch(addEditFolderId(item.key))
@@ -62,8 +60,6 @@ export default React.memo((
 
     function goIntoFolder(e, item) {
         e.stopPropagation()
-        // navigate('/collection/N1') // *****************************************<==== delete when build
-        console.log('enterFolder')
         dispatch(addFetchData({
             item,
             fn: navigate(`/collection/${item.key}`),