+++ src/component/folder/FolderName.jsx
@@ -8,7 +8,6 @@ const Page = styled.div`
     text-align: center;
     width: 84%;
     color: ${({ font }) => font ? 'black' : 'white'};
-    // caret-color: ${({ font, config }) => font ? config.primary : config.tertiary};
     user-select: none;
     outline: none;
     border: none;
@@ -22,7 +21,7 @@ const FolderName = React.memo((
         children,
     }) => {
     const { configuration: config } = useSelector(selectGlobal)
-    console.log('folderName')
+
     return (
         <Page
             config={config}