+++ src/page/collection/Main.jsx
@@ -33,8 +33,6 @@ export default function App() {
     const [saveWarning, setSaveWarning] = useState(false)
     const [hide, setHide] = useState(false)
 
-    console.log(currData);
-
     return (
         <Main config={config}>