+++ src/redux/slice/globalSlice.js
@@ -6,28 +6,20 @@ import { fetchData, setDataToLocal } from "src/utils";
 export const GlobalSlice = createSlice({
     name: 'global',
     initialState: {
-        // configuration: initialConfiguration,
         configuration: await fetchData('Juuten_Configuration', initialConfiguration)
-
     },
     reducers: {
         setConfiguration: (state, action) => {
-            console.log(action.payload)
-
             const newConfig = {
                 ...state.configuration,
                 [action.payload.key]: action.payload.value
             }
-            console.log(newConfig)
-
             setDataToLocal('Juuten_Configuration', newConfig)
             return {
                 ...state,
                 configuration: newConfig
             }
         },
-
-
     }
 })
 //